requirements.txt
pyinstaller>=6.2.0
psutil>=5.9.6
colorlog>=6.8.0
ruamel.yaml>=0.18.5
requests>=2.31.0
tqdm>=4.66.1